use proper uintptr_t casts for pointers

+++ src/cpc/diskimage/dsk.c
@@ -25,6 +25,7 @@
 #include "diskimg.h"
 #include "dsk.h"
 #include "../host.h"
+#include <stdint.h>
 
 static int	DskGetSectorSize(int);
 
@@ -61,7 +62,7 @@ int		Dsk_Validate(const unsigned char *pDiskImage, const unsigned long DiskImage
 					{
 						/* correct size image */
 						int	i;
-						DSKTRACKHEADER *pTrack = (DSKTRACKHEADER *)((int)pHeader + sizeof(DSKHEADER));
+						DSKTRACKHEADER *pTrack = (DSKTRACKHEADER *)((uintptr_t)pHeader + sizeof(DSKHEADER));
 
 						for (i=0; i<TotalTracks; i++)
 						{
@@ -80,7 +81,7 @@ int		Dsk_Validate(const unsigned char *pDiskImage, const unsigned long DiskImage
 							if (thisTrackSize>TrackSize)
 								break;
 
-							pTrack = (DSKTRACKHEADER *)((int)pTrack + TrackSize);
+							pTrack = (DSKTRACKHEADER *)((uintptr_t)pTrack + TrackSize);
 						}
 
 						if (i==TotalTracks)
